django-app/requirements.txt
pymongo==3.7.2
simple-websocket-server==0.4.0
six==1.12.0
tornado==5.1.1
websocket-client==0.54.0
